# Optimize probing tables