--- a/hashprobe/hashprobe_amd64.s
+++ b/hashprobe/hashprobe_amd64.s
@@ -47,6 +47,7 @@ test:
     CMPQ SI, DI
     JNE loop
     MOVQ CX, ret+104(FP)
+    VZEROUPPER
     RET
 insert:
     MOVL R13, R11