… hashprobe/hashprobe_amd64.go
… hashprobe/hashprobe_purego.go
… hashprobe/hashprobe_test.go