Based on #260, this PR modifies the memory layout of probing tables to improve performance of the data structures.

I documented the designs in the code, at a high level the key change is a reduction of the number of CPU cache lines that need to be touched during lookups of keys in hash tables, and using SIMD operations where possible to amortize the cost of lookups.

The change is particularly effective at increasing performance of of lookups in tables with 10-100K keys, as larger tables were more affected cache misses since they exceeded the size of CPU caches. The less cache lines we touch, the lower the cost of memory accesses.

```
name                                                  old time/op  new time/op  delta
Dictionary/Insert/INT32/N=100                          242ns ± 0%   237ns ± 0%    -1.96%  (p=0.000 n=8+10)
Dictionary/Insert/INT32/N=1000                        2.50µs ± 0%  2.15µs ± 0%   -14.04%  (p=0.000 n=9+10)
Dictionary/Insert/INT32/N=10000                       61.9µs ± 1%  24.3µs ± 0%   -60.74%  (p=0.000 n=10+10)
Dictionary/Insert/INT32/N=100000                       696µs ± 1%   361µs ± 6%   -48.19%  (p=0.000 n=9+10)
Dictionary/Insert/INT32/N=1000000                     12.3ms ± 6%   8.3ms ± 9%   -32.20%  (p=0.000 n=10+10)
Dictionary/Insert/INT64/N=100                          256ns ± 0%   255ns ± 0%    -0.47%  (p=0.000 n=10+10)
Dictionary/Insert/INT64/N=1000                        2.85µs ± 0%  2.29µs ± 0%   -19.85%  (p=0.000 n=8+10)
Dictionary/Insert/INT64/N=10000                       65.0µs ± 0%  26.5µs ± 0%   -59.25%  (p=0.000 n=10+9)
Dictionary/Insert/INT64/N=100000                       750µs ± 3%   748µs ± 2%      ~     (p=0.549 n=9+10)
Dictionary/Insert/INT64/N=1000000                     14.7ms ± 1%  12.2ms ± 4%   -16.86%  (p=0.000 n=8+10)
Dictionary/Insert/FIXED_LEN_BYTE_ARRAY(16)/N=100       474ns ± 0%   404ns ± 0%   -14.78%  (p=0.000 n=10+10)
Dictionary/Insert/FIXED_LEN_BYTE_ARRAY(16)/N=1000     5.05µs ± 0%  4.62µs ± 0%    -8.67%  (p=0.000 n=10+9)
Dictionary/Insert/FIXED_LEN_BYTE_ARRAY(16)/N=10000     105µs ± 2%    95µs ± 1%    -9.59%  (p=0.000 n=10+10)
Dictionary/Insert/FIXED_LEN_BYTE_ARRAY(16)/N=100000   1.25ms ± 3%  1.20ms ± 6%    -3.61%  (p=0.001 n=10+10)
Dictionary/Insert/FIXED_LEN_BYTE_ARRAY(16)/N=1000000  29.8ms ± 2%  27.6ms ± 2%    -7.29%  (p=0.000 n=9+9)

name                                                  old value/s  new value/s  delta
Dictionary/Insert/INT32/N=100                           413M ± 0%    422M ± 0%    +2.00%  (p=0.000 n=8+10)
Dictionary/Insert/INT32/N=1000                          400M ± 0%    465M ± 0%   +16.36%  (p=0.000 n=9+9)
Dictionary/Insert/INT32/N=10000                         162M ± 1%    412M ± 0%  +154.69%  (p=0.000 n=10+10)
Dictionary/Insert/INT32/N=100000                        144M ± 1%    278M ± 6%   +93.24%  (p=0.000 n=9+10)
Dictionary/Insert/INT32/N=1000000                      81.6M ± 5%  120.5M ± 9%   +47.69%  (p=0.000 n=10+10)
Dictionary/Insert/INT64/N=100                           391M ± 0%    393M ± 0%    +0.48%  (p=0.000 n=10+10)
Dictionary/Insert/INT64/N=1000                          351M ± 0%    437M ± 0%   +24.77%  (p=0.000 n=8+10)
Dictionary/Insert/INT64/N=10000                         154M ± 0%    378M ± 0%  +145.42%  (p=0.000 n=10+9)
Dictionary/Insert/INT64/N=100000                        133M ± 2%    134M ± 2%      ~     (p=0.549 n=9+10)
Dictionary/Insert/INT64/N=1000000                      68.1M ± 1%   82.0M ± 4%   +20.33%  (p=0.000 n=8+10)
Dictionary/Insert/FIXED_LEN_BYTE_ARRAY(16)/N=100        211M ± 0%    247M ± 0%   +17.34%  (p=0.000 n=10+10)
Dictionary/Insert/FIXED_LEN_BYTE_ARRAY(16)/N=1000       198M ± 0%    217M ± 0%    +9.50%  (p=0.000 n=10+9)
Dictionary/Insert/FIXED_LEN_BYTE_ARRAY(16)/N=10000     95.1M ± 2%  105.2M ± 1%   +10.60%  (p=0.000 n=10+10)
Dictionary/Insert/FIXED_LEN_BYTE_ARRAY(16)/N=100000    80.0M ± 3%   83.1M ± 5%    +3.77%  (p=0.001 n=10+10)
Dictionary/Insert/FIXED_LEN_BYTE_ARRAY(16)/N=1000000   33.6M ± 2%   36.3M ± 2%    +7.85%  (p=0.000 n=9+9)
```